Introduce MigrationManager
