+++ ios/MullvadVPN/SettingsManager/TunnelSettingsV2.swift
@@ -13,18 +13,6 @@ import struct WireGuardKitTypes.IPAddressRange
 import class WireGuardKitTypes.PrivateKey
 import class WireGuardKitTypes.PublicKey
 
-/// Settings and device state schema versions.
-enum SchemaVersion: Int, Equatable {
-    /// Legacy settings format, stored as `TunnelSettingsV1`.
-    case v1 = 1
-
-    /// New settings format, stored as `TunnelSettingsV2`.
-    case v2 = 2
-
-    /// Current schema version.
-    static let current = SchemaVersion.v2
-}
-
 struct TunnelSettingsV2: Codable, Equatable {
     /// Relay constraints.
     var relayConstraints = RelayConstraints()
